+++ configure.py
@@ -890,6 +890,7 @@ scylla_core = (['message/messaging_service.cc',
                 'cql3/attributes.cc',
                 'cql3/cf_name.cc',
                 'cql3/cql3_type.cc',
+                'cql3/description.cc',
                 'cql3/operation.cc',
                 'cql3/index_name.cc',
                 'cql3/keyspace_element_name.cc',
